Add ability to select closest node in follow mode.
I was reluctant of adding this feature as I thought it could be
confusing for new users to select nodes that sometimes were invisible
and you could just switch to edit mode, select the node, and then go
back to follow mode.

The feature proved invaluable in certain situations though, so I've
decided to add it.