Update to Unity 2017.1.
I had to set the Api Compatibility Level to .NET 2.0 instead of .NET 2.0
Subset to be able to build the project in this version of Unity.

I'm not sure if there's a way to revert that change and still have the
project build without problems, but I'd like to do it if possible.