Create auto paste wp to chat.ahk
A script I've been using with Autohotkey that will register if a WP (or any in-game code really) has been copied to the clipboard and paste it to chat. Makes the maprunning experience a lot better since the code from a node that was consumed automatically gets put into chat.