Update project to Unity 2019.2.6f1
I had to include the source for the MumbleLink libraries as one of them
had to be tweaked slightly to fix an bug introduced by the update.

Api Compatibility Level has been bumped to .NET 4.x.

`Use DXGI Flip Model Swapchain for D3D11` option had to be turned off in
Unity player settings for the transparent window to work as before.