Normalize line endings
Line endings for all text files should be LF in the repository, and
they should be able to have any line ending in local working trees.

Exceptions are certain Unity generated files which should be kept as LF
always since they are always generated that way, and json files which
are kept as CRLF so they show up properly on all Windows text editors.